extract Configuration interface
