+++ src/main/java/com/dfsek/tectonic/loading/object/ObjectTemplateLoader.java
@@ -1,7 +1,7 @@
 package com.dfsek.tectonic.loading.object;
 
 import com.dfsek.tectonic.abstraction.TemplateProvider;
-import com.dfsek.tectonic.config.Configuration;
+import com.dfsek.tectonic.config.YamlConfiguration;
 import com.dfsek.tectonic.exception.ConfigException;
 import com.dfsek.tectonic.exception.LoadException;
 import com.dfsek.tectonic.loading.ConfigLoader;
@@ -22,7 +22,7 @@ public class ObjectTemplateLoader<T> implements TypeLoader<T> {
     public T load(AnnotatedType t, Object c, ConfigLoader loader) throws LoadException {
         ObjectTemplate<T> template = provider.getInstance();
         try {
-            loader.load(template, new Configuration((Map<String, Object>) c));
+            loader.load(template, new YamlConfiguration((Map<String, Object>) c));
         } catch(ConfigException e) {
             throw new LoadException("Unable to load object.", e);
         }
