+++ src/test/java/TestObjectLoader.java
@@ -1,4 +1,5 @@
 import com.dfsek.tectonic.config.Configuration;
+import com.dfsek.tectonic.config.YamlConfiguration;
 import com.dfsek.tectonic.exception.LoadException;
 import com.dfsek.tectonic.loading.ConfigLoader;
 import com.dfsek.tectonic.loading.TypeLoader;
@@ -10,7 +11,7 @@ public class TestObjectLoader implements TypeLoader<TestObject> {
     @SuppressWarnings("unchecked")
     @Override
     public TestObject load(AnnotatedType t, Object c, ConfigLoader loader) throws LoadException {
-        Configuration configuration = new Configuration((Map<String, Object>) c);
+        Configuration configuration = new YamlConfiguration((Map<String, Object>) c);
         return new TestObject(loader.loadType(String.class, configuration.get("string")), (int) loader.loadType(int.class, configuration.get("number")));
     }
 }
